bring in latest Makefile, move code to src
