+++ .gitignore
@@ -1,3 +1,4 @@
+.vscode
 .DS_Store
 *.psd
 #*.png
