+++ src/AIMgr.cpp
