+++ src/AIMgr.h
