+++ src/StringInputState.cpp
