+++ src/StringInputState.h
