+++ src/resource.h
